feat(floaty): wait for floating permission granted when creating a window
